requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
tensorflow>=2.8.0
optuna>=3.0.0
featuretools>=1.0.0
flask>=2.0.0
pyyaml>=6.0
scipy>=1.7.0